"1:50pm.
```
let inline response history one two actions next =
    memoize game (fun (history, one, two) ->
        let response = response one actions next
        ordered_game.Add(response)
        let node = memoize agent (fun _ -> node_create actions) (history, state one)
        let node = memoize game_nodes (fun _ -> node) response
        response
        ) (history, one, two)

let inline chance one dice next =
    let chance = chance one dice next
    ordered_game.Add(chance)
    chance
```

This should do it.

3:20pm.

```
// Forget this. This algorithm is too complex for me to deal with. I barely started, and it is already a huge mess.

// What I intended to implement here was FSI-CFR from the paper, but though the pseudo-code is relatively straightforward,
// the part about getting a lexicographically ordered array of game nodes is what is difficult. It is much more difficult
// than the actual algorithm itself.

// CFR has a relatively neat monadic implementation. Turning a game into a graph on the other hand is more like a complicated
// compiler project than anything else. It already exceeds my threeshold for allowable mutation, and I can tell that it is going
// to take me something like a week to get it done correctly.

// The reason why is it worked for the authors is because they hand specialized everything just for the sake of one very specific game.
// In general, while the paper itself is fine, the code examples in it are fairly horrid all around.

// One thing that particularly sucks about FSI is that the actual game graph itself is much larger than the amount of player nodes.
// For Dudo here, what needs to be memoized is the state of the other player (6x) and the player ids (2x).
// So for this that means a 12x increase in memory consumption.

// Since I haven't restricted the length of the history yet here, there would be no benefit to FSI over the standard program, but
// it would be a lot slower as everything would be interpreted and the program would be spread out across memory.

// It is really difficult - for CFR with memoization it would be great if there was a neat way of doing a topologically sorted kind
// of traversal without necessarily serating the game into its graph representation and having to do complex compiler analyses before
// that.

// As I said, just forget FSI. I can only hope that other variants of CFR do not end up being as complex as this.
```

I am throwing in the towel on this. I had enough. I simply do not want to waste another week for the sake of something I do not need.

Let me move on.

I definitely need a break as this thing has so exhausted me. I hate the paper now. I wanted to clear this and then move on with a victory behind me, but now I will have to slink off in defeat. I was thinking of doing a writeup on CFR and posting it on the RL sub, but now it would be embarassing.

But...

There is a beginning of an idea that I had earlier but could not quite work out. It might be possible to mode the graph as a network of asynchronous agents. How to handle compilation for that is anybody's guess. I have no idea at the moment."