"8:45am. Let me go down this route. Had I already known DT because I started in 2007, I would be so much ahead now. Let me prioritize leveling up my programming skills for the time being.
It is time for the Little Typer.

9:05am. Focus, me focus.

9:35am. I am really, really hungry. I hadn't had enough to eat yesterday.

Let me have breakfast now.

10:20am. I had a hearty meal. Now that I am no longer starving, let me resume the book. I read the postmodernism slides on the /ratanon/ board.

Hmmm...let me say a few things about it then. The thread of reasoning spun throughout that material is very modern in character - for example it gives out the names of a bunch of philosophers as if that means anything. Postmodernism which I will define as the current era for clarity is as much about shedding than about building on what came before.

There isn't much in the old schools worth building upon, so they might as well be set on fire and burned down to the ground to make way for better things.

By all appearances the defining trait of rationalism is the fetishistic lack of understanding of desire. In that sense those slides fit that mindset to a T. I am not sure the author would identify himself as a rationalist, but the situation would resemble an atheist not identifying himself as a Christian. It makes little difference to shed aside the rituals if one is going to keep the morals and the patterns of thinking to those in opposition.

10:35am. We killed the old gods like Zeus and made them our power. The natural elements were stripped of their godhood and remade into technology.

The next step would be to kill the big one. Kill the man behind the God so as to leave only the power behind. Harness the essence of perfection to enter the loop and elevate the indivisible desire. That is the way to go.

This world is not letting me proceed as planned, but I made up my mind over 4 years ago. The core of my identity is the desire. I do not have the courage nor the power. Despite all the failures, all I can do is to continue down the path of dreams.

And so I shall."