"1:35pm. Let me resume. I still have a few good hours left in the day.
1:55pm. Ah, I think I understand how to implement primitive fold combinator without a fixpoint operator. I get it now.

2:15pm. 137/418. I am a bit confused as to how rec-List is actually implemented...

Ah, no wait, the end is the state here. I am such a goof.

2:20pm. Focus me.

2:30pm. 140/418. I am skimming this stuff, as I am not too interested in reasoning about List operations. I went through this hundreds of times already in different contexts. The thing that is interesting is the passing of first class types.

2:35pm. 144/418. This seems like it will have promise.

2:55pm. 151/418. Ah, I should have seen this coming. Quite brilliant. So there is more to forall that meets the eye.

3:50pm. 167/418. By this point my focus is really starting to dim.

4:10pm. Let me stop here for the day. I'll resume from 164/418 tomorrow. I just can't go on any longer.

At the start of the day, I was not sure whether to go with this or CFR, but it feels like I've made the right decision here.

This here is real programming with dependent types. That Coq stuff was complete nonsense after all. I can make do with this. Function application is something that I know and love. It is what I am familiar with.

4:25pm. Let me go do the chores while Kotobuki downloads. It is time to rest. Maybe I'll start surfing the types subs."