"10:35am. 153/418. I am reviewing these sections again. So the pi bindings are actually one level lower on the type hierarchy than their annotations? Makes sense.
158/418. Let me go through this chapter from the top. After that I'll take a break. There are some chores that need doing.

11:20am. 172/418. Ok, by this point even I am starting to tune out and not because I am tired. This sort of thing is what I should be following along on the computer rather than mentally.

"Getting the right answer is worthless if we do not know that it is correct. Understanding the answer is at least as important as having the correct answer."

The book says some things I am forced to agree with. One does not make a big deal of it usually, but decades of ML really hammers in the lesson of this home.

11:30am. 179/418. Let me stop here for now.

I am going to have to redo this chapter. I'll probably find the Pie language in Racket somewhere. I need to go through the chapter using it as I can no longer keep track of all the stuff in my head."