"1:15pm. Got Pie in from the Racket package repository.
```
(claim qwe (-> Nat Nat Nat))
(define qwe (λ (a b) b))
```

Ok, this works. Though it seems + is not in the language.

...This is going to make things complicated. Do I have to do the entire book first?

1:20pm. http://the-little-prover.github.io/

1:25pm. Ok, I know how to do lambda in Racket, but how do I do pi?

https://docs.racket-lang.org/pie/index.html

Ah, there are docs for it.

1:35pm. https://www.reddit.com/r/programming/comments/a2s08c/the_little_typer_lets_learn_about_dependent_types/

1:45pm.

```
#lang pie

(claim qwe (Pi (A U) (-> Nat Nat Nat)))
(define qwe (λ (A a b) b))
```

There is some kind of error here that I can't figure out.

1:45pm. https://corecursive.com/023-little-typer-and-pie-language/?src=reddit

I'll listen to this podcast with the authors later.

```
(claim qwe (Pi ((A U)) (-> Nat Nat Nat)))
(define qwe (λ (A a b) b))
```

Ah, it seems Pi takes a list of lists. Ok.

2:05pm. Ok, I've gone over the reference.

There is so much in the Pie language I've yet to understand. It seems I only ever scratched the surface of dependent typing with Spiral.

2:10pm. Let me go through chapter again. This time I will write the examples myself into the script."