"https://boards.4channel.org/a/thread/186240725#p186252560 isao.png "What was his name again?" "Enola Great." "Isao "They can't retreat if you melt the streets" Harris." ""AnCaps in town? Burn it all down" Isao" ""Dresden was an appetizer" Isao" "Incendiary Isao." "Magic Man "Pop a cap into an ANCAP" Isao"
1:40pm. Let me finally resume.

I need to focus on the example that I did and train myself so I can trace the flow in my mind. Once that is done I should come to some sort of an understanding of things.

1:45pm. Let me commit here.

I need to rebuild my mental models from the ground up. Let me once more check the language reference."