"5:45pm. I didn't really think much about `replace` while in bed, but I did reflect a lot on my goals. I think since things are like this that I am just going to dedicate the next few years to math through dependently typed programming. I need to gain the ability to say intelligent things about learning algorithms and not just implement them. Benchmarks do not mean anything. They would if the program was simple, but these certainly aren't.
For all I know, the algorithm I am looking for might perform much worse than backprop initially before its performance gets good. Benchmarks of the kind that are being done right now would completely hide that. They will always tilt the field towards domain knowledge.

What kind of domain knowledge is embedded in CFR? Why is it a good algorithm? For CSI remember that hunch that I had that the stuff in the paper is wrong? How do I really get to the essence of that?

Right now all I have is a bunch of ideas that I can't make anything of.

The lack of formalism is killing me. I can't possibly win like this. Once again, what I want to do and what I need to do is completely different.

5:55pm. Even so, I can't just become a pure mathematician. The only way I can evolve from here is to bridge math and programming. Terrence Tao does not use computers, but I've always been bad at math. I will never get anything without computers. Nor will I ever attain anything of value without programming.

Since the Pharo move failed, there is literally only a single path available to me.

CFR is just a distraction. When I decided in January that I no longer want to test that exploration method, I've failed the quest right then and there. It just took me a while to realize that. I've failed much earlier, even before I began.

I am too green after all. I've flattered myself by thinking I could be a good programmer, but bad mathematician. The truth is being bad at either of those things means that you are bad at both. The purpose of all of this is to cause the Singularity. It is impossible to cause it without superb scores in both.

6pm. I've made up my mind. Forget poker. Maybe at some point I will get a chance to try it again when I find that algorithm that I am seeking.

I made the right choice by picking Little Typer. Grasping the fundamentals of mathematics and making it one with programming is what I should be doing. It will be what I will be doing for the next few years.

And if this is right, then obviously everybody else in the field of ML is doing it wrong. I won't accept any other conclusion. If using the machines to do math is right, then it means that those who aren't are in the wrong.

That will give me an edge that I seek.

6:05pm. One thing that is really cool in DP is how the types are program generation templates. There should be links to ML there. With theoretic breakthroughs it might be that I will be able to use machines to generate algorithms themselves given a specification in types.

But damn it, I seriously am never going to make money in the way I would prefer will I? I should have known, my money making expectations just keep getting betrayed. But it is fine. This world is my punishment. It is a prison that I must break free.

6:10pm. Now let me fiddle a bit with replace. I really did not do much today, but I still want to continue."