"6:15pm. https://docs.racket-lang.org/pie/index.html#%28form._%28%28lib._pie%2Fmain..rkt%29._~ce~a3%29%29
Ok, I see. The output does not have to be an equality type necessarily.

6:45pm. Done with lunch. I've realized that I cannot really implement `cong` with replace in any straightforward manner. If I could somehow rip the `from` field from the equality and term cast it then I could do it, but I do not have access to that just yet.

6:50pm. The type annotation for replace are ambiguous. That `X` should really be an `U` in the `=` constructor. It should be something like `Pi ((X U))`, but nevermind that.

For once, now that I know how, instead of fiddling with it just like I did in Coq hoping to get a match, let me see if I can reason out what the solution for `twice=double` should be.

I've already figured out the base.

The step should be a replace.

The base of the replace should just be `twice n-1`. This should match the `motive from`. This should be easy.

Now comes the part that confused me earlier, the motive.

Really, the important part is not the induction. The important part in doing the replace is that I can get the `n: Nat` which I need to do the replace.

Once I have that, all I need to focus on is making the left side match the right.

Easy as pie. Now why is this giving me so much trouble? Because I've lost sight of the big picture.

Somehow while doing Coq what induction meant to me was somehow adding a number to a list. Like, I should have `List n`, and then I would have `List (n+1)` after that or something. I must have been braindead. What is important is that I have the number itself!

Now let me do this thing."