"7:05pm.
```
rec-Nat is primitive recursion on Nat. If target is zero, then the whole expression is base. If target is (add1 n), then the whole expression is (step n (rec-Nat n base step)).
```

I've done this a lot, but it still trips me up than n is n-1 here. What the hell?

7:10pm. No, forget it. It feels like I've been only just now been able to regain some of the feeling I get when I usually program. That having said, I am still dead tired at this point in the day. I do not have the energy to bring the problem to a close right now.

7:15pm. I will stop here for the day. I will be able to make better progress tomorrow. Today I just had the first real thoughts on what proving really is. I might go through the book from the top tomorrow and do the proofs with my own hand without looking at the solutions. I should be able to do this now."