Let me finally start the day. I've been making good progress through the book so far. The only area where I am particularly weak in are those extrinsic equality proofs that require partial evaluation of expressions.
I do not have the mental machinery to do those properly, but with some practice I will be able to get myself to the right level.

As I thought, Spiral was a good stepping stone for real DT programming. In terms of expressiveness, Spiral is somewhere in-between F# and DT languages. Going from F# to Coq and Idris was too much for me, but going from Spiral to Pie is just right.

I need to get better. I still have room to improve.

Yeah, Pharo-like platforms and visualizations are important. Even if not Pharo, the next ML is going to have to provide such an experience.

But right now, the far and the biggest number one concern is the lack of good ML algorithms. Everything hinges on this one thing.

If I can integrate programming with math and unlock a new ability tree for myself, that might open the road that lies ahead. Since the entire field of ML is incompetent, I can only rely on myself.

For what lie ahead, I must find the algorithms that I need on my own. I must make my own strides in understanding credit assignment. If I am really serious about this, I cannot back down.

I need to keep going forward.

It is time that I resume the book. Only a 100 pages are left.