"11am.
```
What I have a need for is to integrate all my experiences into one.

If I were a sane, sensible kind of person I would not try making my own library in F# back in 2016. And then once I got stuck at that, I would not spend two years working on [Spiral](https://github.com/mrakgr/The-Spiral-Language). I would have gone with Python back in 2015 and spared myself the work.

As I've worked, I also kept getting better which served as the fuel for the creation of Spiral.

I think I could definitely move from F# + Spiral to something else, but that cannot possibly be an impoverished language like Pharo. If I were the person that could do this, then I would probably never have started work on Spiral in the first place and kept at it for so long.

Since trying to get better at ML by focusing on visualizations is a dead end that leaves only one thing left. It is going to be a pain and I'd rather do it in some other life, but I just have no choice at this point. I can't even imagine how much this is going to shave off my life, but I have to do it. It is going to be a lot harder than porting Spiral to Pharo and making a visualization library would have been.

I am going to learn dependently typed programming and all that entails including proofs and the general mathematical reasoning which comes as a part of the package. I've tried this in 2016 and did not like it at all, but now I have an actual reason and two years of programming in Spiral to serve as the backbone. My Spiral experience did give me some insight into what programming with first class types entails, though the feedback that I got from the [Little Typer](http://thelittletyper.com/) book shows me that I still hillariously bad at proofs. I am quite good as could be expected in reasoning about what happens when values are propagated forward, but a lot of this stuff goes quite beyond that.

My motivation since 2016 was just to make a ML library and leave the strenous work of figuring out the right algorithm to the community. Three years later the library is here, but the algorithm is not. Instead there are thousands of useless papers with incremental gains to sift through and no real insight. It is quite unnatural that for all the gains deep RL had like in Starcraft and Dota, those same algorithms still crash and burn on toy tasks.

I was more than happy to allocate respects while it looked like they would deliver something useful, but right now I feel they are just padding their resumes and cashing in on the hype. I cannot [trust them](https://www.reddit.com/r/MachineLearning/comments/azvbmn/n_openai_lp/).

There has to be an algorithm that is more roboust and scalable than backprop, and I think it is an embarassment that the entire ML community cannot muster better criticism against it than making vague analogies to the human brain. It is known for a fact that Bayesian algorithms like [PILCO](http://mlg.eng.cam.ac.uk/pilco/) can exceed the sample efficiency of deep learning by orders of magnitude. If [Bayesian reasoning](http://probmods.org/) is a template for how learning should be done, then deep learning does not fit the mold.

I've decided that I cannot leave the future of AI to mathematicians.

If it is merely for the task of making a poker agent, I could do something such as implementing a [counterfactual regret](https://github.com/mrakgr/The-Spiral-Language/tree/a40eae68990a535d6c766483f72d579d01941468/Learning/CFR/Intro/CFR) based agent, but that is merely a compromise. It would distract me from what I would need to do afterwards.

As much as Spiral's purpose was to make a ML library, it was also a hedge against the possibility that deep learning might not end up being the solution. If deep learning falls out of favor, frameworks like TF will die out, but Spiral will be just as great as ever.

It is not a given that I cannot do it even though I've historically been bad at math. If I can make math into programming I could do it.

The aim behind learning dependently typed programming is simple - if I could give abstract enough types to various existing algorithms that will give me great insight into their general family - into their properties and their flaws. Not only would that allow me to derive better general algorithms, but also to possibly specialize them to a given domain.

Having good enough types would open the door not to just implementing them, but also to potentially synthesizing them. It might be possible deep into that process to put such general ML algorithms into the compiler and have them augment the synthesis process itself. This is merely taking the program synthesis vision of dependent types to its logical conclussion.

Right now, from my current perspective, the process of typing looks a bit like the process of deoptimization. It makes sense if you think about it - to make a program run fast, one needs to get rid of all the abstraction and the scaffolding that provides the higher level structure. It is not trivial to go the other way around. At the assembly level it would take a massive amount of effort to reengineer even trivial programs.

I am certain that all the ML algorithms that are currently in existence whether it be the tabular CFR or the deep learning ones are merely optimized forms of some higher order abstract structures.

If this vision is right then everything else is wrong.

I am wrong. It is an absurdity to see myself as good at programming and bad at math. Isn't it instead the case that I am bad at it period if programming and math are equivalent?

The ML community is wrong. Mathematical insights being separate from algorithmic implementations is a huge source of friction. It is ridiculous that people rely on other people for feedback on insights rather than machines. It is disgusting. Making math less of a social activity should be one driving goals of the quest for intelligence.

I need a better reason than pride or money, a better unified language, and a better stable and versatile platform.

I've lost, but since I want to keep playing I've no choice but to restart and try again. That is what the games have taught me. I need to belive in the luciferian mistrust called skepticism.

The thing I need to believe in the most is that the skills I will gain from the retraining will be worth gaining.
```

My masterpiece is done. I've finally said it - there is no point in being good at programming and bad at math in this world. In a way the last four years were just the prologue to this. From April 2019, it will be the real thing.

11:10am. Now, what's for breakfast?"