I wonder if I should keep making these empty commits since I will be doing Idris and then dependently typed programming for the time being?
I'll defer the decision till the end of the day.

Forget about that for now. Let me resume the book.

I've been thinking about how one might effectively compile dependently typed programs and cannot come to anything definite. Inductive datatypes are a whole another ballgame from anything else.

The techniques I am using in Spiral are not enough.

But ultimately, that is not important at the moment.

I only started work on Spiral after over a year of experience in F#. Effectively compiling dependently typed programs is a question that can be answered later.

Dependently typed programming is what comes now.

The only benefit that I've gained for my 4 years is the desire to trully attempt to solve the credit assignment problem on my own.

For somebody as lacking of talent as myself, I am going to need the full power of machines to do this.

With Spiral, I solved primitive compilation issues with ML libraries, but that is not what I need.

I need the right algorithm. And to find it I need the power to move about in the darkness without stumbling.

To do this I need a new style of programming. I need to reach the pinnacle.

I need to leave nothing unturned.