Today I got up literally at nearly 11am so I missed most of the morning. But my rule is to not regret sleep under any circumstances.
I am done with breakfast and have the next 5 hours left for the Idris book.

I will start the day by going over some of the Pie examples and then I will do the exercises from the book. After that I will be done with chapter 7.

I'll aim to get a few more chapters packed away today and hopefully I should be done with it by the end of the week.

At the rate I am going, a few months of exercise with dependent types should get me to proficient level with it.

Today is another step in that journey through the long desert. Let me start.