Today I will go through the Streams chapter of the book.
Yesterday I gave the Demonsteele Doom mod a try and ended up playing till 11:30pm. Good times were had. I have to thank the anon who mentioned it in one of the Asuka threads - I was not the one talking to him, but I decided to give it a try anyway.

I have one good hour before I take a real break so let me make use of it.