"9:45am. Let me finally start with the Idris book.
10:20am. 386/484. How many pages till the end of the chapter? 13.

Well, this is not too bad. I should be able to do this by breakfast time.

10:35am.

```
        Vend
        Display "Enjoy!"
        machineLoop
```

Let me back this up here.

10:50am. 389/484. Focus me. Stop surfing the PL sub. 10 more pages to go. I can do this before it is time for breakfast.

I'll skip the exercises here.

11:30am. Focus me. Finish the damn chapter. Stop surfing /a/.

11:45am. 399/484. Going to skip the exercises for 13 as they are pretty trivial. Finally at chapter 14.

I am going to finish 14 today and then the book tomorrow.

This is an interesting demonstration (I am yawning here) of what an advanced type system can do, but compiling this is going to be a bitch. I can tell.

At most, all I can hope is that DP programming gives me a hint of how to do ML better. I want to go through Coq and study type theory more in depth before I come to a conclusion.

11:50am. Let me finally stop here for the morning."