"12:15pm. Focus me, focus.
1:05pm. Damn it is taking long to typecheck even on these trivial 100 liners.

1:25pm. 453/484. I am so close to the end now. Just two examples to go through and then I am done with the TDD in Idris book.

1:50pm. 456/484 Focus me. Just 3 more pages and then I can take a break.

2:10pm. Yessss. YESSSSS! I am done with the book!

http://www.cs.nott.ac.uk/~psztxa/publ/ydtm.pdf
Why Dependent Types Matter

Let me read this paper. Epigram was the predecesor to Agda I think.

After that I'll ease myself into the category theory book that I got. I'll also look for Software Foundation supplements. I never looked for it before, but maybe there are some courses that use it to teach online. Maybe I'll find something on Youtube.

2:35pm. https://www.youtube.com/playlist?list=PLGCr8P_YncjUT7gXUVJWSoefQ40gTOz89
Benjamin Pierce - Software foundations

As I expected there definitely was something on this. The playlist is shorter than I thought it would be though.

2:45pm. Ah, forget the paper. Let me just do the book for a bit.

2:50pm. "In this personal note, we ask that readers try to use what they learn in this book to do something they would call “good,” in terms of contributing to the society they’d want to live in. For example, if you’re planning to study this material with others, consider specifically inviting someone from an under-represented minority—a group that is more highly represented in society than in upper-level math classes to your study group. As another example, perhaps you can use the material in this book to design software that helps people relate to and align with each other. What’s the mathematics of a well-functioning society?"

Don't worry, I will be sure to do "good" with the material from the book assuming anything in it turns out to actually be useful.

This timeline really sucks. Why aren't scientists seekers of omnipotence and masters of the dark arts instead of just a bunch pussies? I'd enjoy such a world much more than the one I am in right now. Allowing normies to rule the world was a mistake.

But to be fair, science is too much tilted towards augmenting society rather than augmenting the individual actor. Right now I can imagine myself having an IQ of 300 and being a master of all the fields of science and it would still be useless as there is no way to really convert that mastery into killing power.

Without a doubt, the only choice is to bring the long path that that technology has taken to its convergence and start the Singularity. I just need a sliver of insight more with regards to how it is possible to do credit assignment. Just a bit more, to make it all come together.

3:05pm. I will continue the long fight. I once swore to myself that I'd do anything to get closer to omnipotence - it would be an embarrassment if I stopped short of learning mathematics. When I swore anything I really meant it."