"3:30pm. "Friendship is not transitive—my friend’s friend is not necessarily my friend—but possible communication of a concept or a disease is."
Interesting that they picked memes as positive and friendship as a negative example of non-transitivity. I can't help but wonder - should friendship really be like this? Friendship today is less of a close connection these days and more like being an acquaintance. I can't imagine this view being a constant throughout history.

Non-transitive links are weak, much like shallow ML models. If friendship cannot be a cause of unity then other lines will necessarily have to take its place.

...This math book is making me think too much along lines I really should not be exercising while studying math.

3:50pm. 20/353. Ugh, I already think I may have made a mistake. I'll go through at least the first chapter. After that I'll take a look at the Coq videos.

3:55pm. Forget this book. I can't internalize this kind of thinking. I guess hard math is its own thing after all.

Let me go through the first Coq video and then I will do the chores and call it a day.

4:05pm. https://youtu.be/KKrD4JcfW90?list=PLGCr8P_YncjUT7gXUVJWSoefQ40gTOz89&t=242

4 minutes in so far and the background on this is interesting. Yeah, this will definitely be more up my alley.

https://www.cs.uoregon.edu/research/summerschool/summer12/curriculum.html

In this summer school there is a lot of stuff here.

4:10pm. https://www.youtube.com/channel/UCDe6N9R7U-RYWA57wzJQ2SQ/search?query=harper

As can be expected, there is newer stuff by Harper in OPLSS channel. Oh, there is even stuff by Brady.

Let me go back to watching the video by Pierce.

4:30pm. Had to take a break. Let me finally watch the thing. Tomorrow, it seems I will go back to what I was doing back in January. Back then I abandoned the attempt due to importing issues. I will have to fiddle around a bit with system paths in order to get that to work.

4:45pm. https://www.youtube.com/results?search_query=Category+theory+foundations&page=&utm_source=opensearch

I want to give CT a try again. I want to see if this goes anywhere.

https://youtu.be/KKrD4JcfW90?list=PLGCr8P_YncjUT7gXUVJWSoefQ40gTOz89&t=1150

Let me pause the video here so I can do that. Let me watch the CT for 10m and after that I will take off for the day.

https://youtu.be/ZKmodCApZwk?t=36
"Category theory is, as a first approximation, an abstract algebra of functions."

Yeah, I think the book should have started with this.

"We are looking at the same thing for two different points of view."

Yeah, this is what I wanted the book to start with - on the connection between lambda calculus and category theory.

"It has kind of a geometric character."

This brings to mind information geometry.

Ok, I am sold on at least watching the first lecture of this.

5pm. https://youtu.be/ZKmodCApZwk?t=672

Let me stop for a bit here. I'll do the chores and watch this for another 20m.

5:25pm. Let me resume. Just a bit more and then I am done for the day.

5:45pm. No that lecture is the same abstract garbage as the book. I'll stick with programming and lambda calculus for the time being.

That having said, I'll make attempt type theory too. Now that I've gone through LT and TTD in Idris I wonder if Robert Harper's lectures will make more sense.

Only one way to find out.

Now...Doom. Speed of Doom + Demonsteele is like crack. That Asuka thread anon changed my life. Let me go for it for a bit. I've certainly earned it."