"10:35am. https://youtu.be/LE0SSLizYUI?t=510
The premise of his talk is quite interesting. He says that beginners to type theory are in good shape.

10:45am. https://youtu.be/LE0SSLizYUI?t=1064
"Everyone understand the idea of an algorithm."

10:50am. https://youtu.be/LE0SSLizYUI?t=1330

Well, yes, this is surprising. This is exactly how Spiral does it.

10:55am. https://youtu.be/LE0SSLizYUI?t=1459

Oh, this last part is where it deviates from Spiral.

11:10am. https://youtu.be/LE0SSLizYUI?t=2269

"The situation about what is true has a much higher and in fact unbounded quantifier complexity whereas the question about what is formally written down in any formalism is always relentlessly recursively innumerable. Which means that the things you write down as a formalism can never come anywhere near what is true. That is Godel's theorem."

For once I can follow this somewhat. Yeah, this is programming stuff.

I can tell - I have no all the mental machinery that I need to grok type theory. Let me go for it then. It is really strange that I find myself engrossed into a lecture such as this.

11:20am. https://youtu.be/LE0SSLizYUI?t=2643

I am a bit confused as to how to interpret this. Ok, I can imagine `Nat/2` being a type, but even so, how would `2 = 4` be true even if that was the case?

...Wait, he said `Nat mod 2`. Perhaps he mean to write `Nat%2`? In that case it would make sense.

Yeah, this is interesting stuff. So types can be used like that.

12:10pm. https://www.youtube.com/watch?v=1U4w0159-Ls

I'll watch the second video later. Let me stop here for now.

12:50pm. Let me resume. It is time for another lecture of this."