"1:30pm. https://youtu.be/1U4w0159-Ls?t=1313
This stuff is what I reasoned about in the v0.1 of Spiral.

2pm. https://youtu.be/1U4w0159-Ls?t=2725

This is all familiar to me from my work on Spiral.

2:05pm. https://youtu.be/1U4w0159-Ls?t=2876

Yeah, this is what I went with in Spiral...

No, actually it is a bit different here though. Spiral would evaluate all the cases unless deferred explicitly.

2:45pm. https://youtu.be/1U4w0159-Ls?t=2036

I am done with the second lecture, but I am reviewing here. This stuff here is what I did in Pie.

Given the nature of this stuff, my focus is the highest it has ever been in my life. It is amazing that I am able to maintain it in face of material such as this. It is making me think that maybe once I am done with this, I should give category theory a proper try - meaning watching the lecture in the morning when I am fresh rather than at the end of the day.

Ok, let me resume the review for a bit and then I will go for lecture 3. Then I'll stop for the day.

2:50pm. Focus me, stop getting lost in thought.

4:35pm. https://www.youtube.com/watch?v=pfOQ97iCIsk&list=PL0DsGHMPLUWXXA8RHzVZ2B5E5hP8CD15Z&index=4

Let me stop here.

4:40pm. I need to take some time off to do the chores. These videos are nowhere as bad as there were years ago. I might be able to fit one more today, but I should try not to push myself. And neither should I try to fit CT or any other kind of learning during my rest hours.

4:50pm. Yeah, I am done for the day. Let me move to the fun stuff. 5h of work is perfectly fine. Working for more than 50% of the time one is awake is just being a slave."