"8:35am. https://youtu.be/pfOQ97iCIsk?list=PL0DsGHMPLUWXXA8RHzVZ2B5E5hP8CD15Z&t=356
I've been following the first 3 lectures, but this stuff is losing me. But I'll stick to the program for today.

9:50am. Done with the first lecture. Now that I've taken a tiny break let me get on with the second one. This stuff is really losing me, I guess to understand this a lot of prerequisite structure is in fact necessary. Thankfully only one lecture is left. After that comes breakfast and then two CT lectures. I'll call it a day after that.

The Coq videogame (Harper called it that in the first lecture) will come after that. And after that...I am leaning towards learning statistics.

How do I put it - I have a quiet determination burning inside of me. It is not quite anger, it is more like injured pride. I just can't believe that I can say so little about the nature of learning itself. It is unacceptable. There is something at the core of this all that I am missing and I do not know what it is.

I've made a language and studied deep learning, but I really do feel looking more into classical methods.

Deep learning drew my in due to its potential, but given that I am having trouble realizing it the fundamentals have gained an allure for me. These PL studies that I am doing right now are just to round off up as a programmer and give me some extra tools that I am going to need when the next part comes.

10am. Well let me start this thing. The sooner I do that, the sooner it will be over."