"12:30pm. Ok, restart this. I think I'll manage to get through 50 pages per day of this.
1pm. 35/748.

The explanations for Bayes rule always go like this and they always feel like gibberish to me. Thankfully I have the experience of going through ProbMods and some understanding of probabilistic programming so now I can put this into the broader context. This will allow me to make headway.

1:20pm. 36/748. When it suddenly switches gears from the toy example into highly abstract nonsense has always taken me by surprise in the past.

1:40pm. 38/748. I am having to translate this into PP concepts in my head, but this makes sense to me now.

I can do this. I think I am finally at a level where a book like this is something I can reasonably digest.

1:40pm. Now if only I can stop being lost in thought, I will be able to make progress through this.

1:55pm. Focus me, focus. I need to get through this, not surf 4chan on the side. Just two more hours of this and then I will take a break. Maybe an early one for the day.

2pm. 40/748. It always gets me how in that course that I watched for example the guy outright states that the likelihood function is not actually a probability. The whole field of mathematics has issue with notation and would do well to move to lambda calculus.

The whole thing makes much more sense once likelihoods are replaced with potentials.

2:15pm. 43/748. While doing research on RL exploration methods back in 2018. I remember once having to ask about the variance of a probability distribution on the math SO. It is right here on this page.

2:30pm. 45/748. This is actually interesting stuff here. Of course I'd skip it the first time around, but now I see its value. If taking the derivative to get the variance directly underestimates the bias in this simple example then of course that would have implications when applying the same procedure elsewhere such as NNs.

No actually, taking the derivative would include some extra terms rather than doing the minization...

I guess figuring out how to do this minization rather than eyeballing it would be one of the things I need to learn.

3:10pm. https://generalizederror.github.io/My-Machine-Learning-Research-Jobhunt/

Let me read this.

3:20pm. Ok, focus me. Let me resume. I should at least finish the first chapter today.

3:25pm. 49/748. These formulas are a bit more complicated and I am not quite sure what they are supposed to be.

3:45pm. "Thus p(r)δr is the probability mass inside a thin shell of thickness δr located at radius r. This distribution is plotted, for various values of D, in Figure 1.23, and we see that for large D the probability mass of the Gaussian is concentrated in a thin shell."

This is what the STAN guy talked about. It is right here in the first chapter.

4:45pm. 60/748. Let me stop here. My focus is completely shot right now. I am getting lost in daydreams rather than reading the book. Somehow thinking about the right algorithm for intelligence is more fun than actually studying the things I will need to make it.

But I'll get into it eventually. This exercise of putting things together in my mind is necessary for motivation.

There are threads running throughout all of this and I need to tie them together.

I need to decide what needs to be kept abstract and what needs to be kept concrete. The failure of deep RL is related to the fact that it is trying to operate on what should be abstractions directly.

Metalearning, it all comes down to that. If only I could figure out a principle for propagating through just one layer of nonlinearity without relying on backprop, that would be a massive breakthrough. Of course, theere are algorithms that do that, but they are all worse than backprop which is the issue. Once you move away from differentiability, then everything become 100 times more complicated.

I understood this well enough which is why I really wanted deep RL to work in its current incarnation.

...I guess the issue is trust. To do credit assignment properly even for a single example, rather than just a single pass like backprop does, a proper Bayesian update would necessitate back and forth communication between the layers to attain convergence.

How could global cohesion be maintained in the face of the fact that the upper layers would not necessarily have connections to sensory information? Meaning, it needs to trust the layers below it.

Should lower layers be given free reign then they will necessarily just feed the inputs that the upper layers want.

These issues of credit assignment come up again and again whether it be on the planetary scale or as low as individual neurons.

Maintaining unity, managing feedback and managing global cohesion is the same thing. This needs to be done in the face of incompleteness.

5pm. I need to think about this more. And as the proper starting point for this sort of hard problem, I need to get the fundamentals of ML off my plate. I did DT programming, which I could not in 2016. I will get this done as well. I certainly do understand the material I have been reading today even though it is quite difficult.

I will call it a day here. Time for rest."